Added super simple test runner.
You can run "make test" to run the suite, which will build stutter
before running tests. If tests fail, a diff will be displayed.

The number of failed tests is returned as the exit code, meaning
that 0 is success, and anything else is failure. The exit code
may differ from the actual number of failed tests if they are
> 255 (in which case there are bigger problems).

Eventually when gc is implemented, valgrind can be integrated here.