Added GC framework and valgrind support in tests.
Added garbage-collection-managed environments and stutter
objects. Basically there is a big doubly-linked list of all
gc objects (of either type), which is managed in each type's
constructor. This allows us to not care about manually freeing
objects in the code, and still let us free everything properly.

The cleanup is now only done on shutdown at the moment. Mark
and sweep collection will be done with this framework in the
future.

The test suite now uses valgrind by default, see readme for how
to disable it if you don't want to use it.