Re-implemented hash functionality
Previous hash implementation had some drawbacks:

1. It was overly complicated due to linear probing
2. It didn't cleanly allow for valid NULL values
3. Returning the keys/values was inefficient

This new implementation addresses the above. In a nutshell, it is
a standard hash with prime number of buckets and uses linked lists
for collision handling.

There is also another doubly-linked list super-imposed on top of all
objects so that the head may be returned as essentially an "iterator"
of the current key/value pairs in the hash. This is also used for
efficient hash destruction.