Change nil output to be printed as () and not nil.
Things are confusing otherwise, because the symbol nil is printed
as nil, but is not equal to nil.
If we want the old behaviour, let's go all the way and make the symbol
nil also equal to nil as it is in Common Lisp.
I favor the current approach (after this commit).