Remove flagging of symbols that correspond to special forms.
That a symbol corresponds to an special form doesn't make a difference
outside of eval. They don't need to carry that information around.
We explicitly initialize them so that they always get the same id as we
were doing with nil and use that information during evaluation.

The next step is to remove the special form functions from the builtins
and move them to the "core."