Refactor _try_ functionality into separate module.
Plus make the base _try_ take a C function pointer catch handler, allowing
the same implementation to work with both C function pointers as well as
stu lambdas (eg by applying a stu lambda with the exception as an argument).

So *Try_eval_stu_catch* is exported to stu and it calls *Try_eval* internally
with it's own custom C catch handler that applies the supplied lambda over
the supplied exception.