Implement Try_eval_list function in try module.
This change is in preparation for a default catch top-level handler in
the public API, as all api functions eval lists of forms as opposed to
just one. Furthermore, all public functions assume that the main env is
passed between calls, hence the "Sv_eval_list_main_env" function which
implements this behaviour.