Make "def" not put directly into the main_env.
In a step towards a module system, this commit makes the "def" builtin not
put bindings directly into the main_env, but instead "capture" the binding
in the interpreter object. This "capturing" is a way to eval a file and
capture all def'd forms without modifying the top level main environment
branch.