Implement scoped "def" & "defmacro".
This patch changes the way things are installed in the environment
in a couple of ways:

  * Anything def'ed in a file is available for subsequent forms in
    the same file, but not between files.

  * Calling "def" or "defmacro" from within a lambda scopes the
    installed key/value pairs to the lambda's scope.

The above means that in order to include the standard library,
for example, it is no longer enough to simply eval the stdlib.stu
file before eval'ing other files or using the repl. In the meantime
the relevant tests dependent on stdlib.stu have been flagged as
"expected failures".

This means that in order for things to be imported between files
(and file-like evaluations) we need a module system!