Add support for structures and opening them.
Field accessor is ::. The field is a static symbol.
Field lookup is currently linear, but that will be addressed.
Defining a new structure is done with deftype.
deftype is a special form that binds the constructor in the current scope
to the symbol that names the type.
Structures are generative for now, so defining a structure type always
creates a new distinct type.
opening a structure means putting binding its fields/values in the
current scope so that they can be accessed without having to refer to
the structure value itself.