Add new defmod & export special forms.
The "defmod" keyword allows for a module name (ie it's type name) as well
as a documentation string to be set. This is meant to be called once per
module, and can be located anywhere within a file.

The "export" keyword allows a symbol to be exported into the the final
module structure, as well as specifying documentation for the export.

This also includes a change that every module form must be explicitly
exported via "export".